+++ DirectFire/Converter/generators/ciscoasa.py
@@ -2,6 +2,10 @@
 
 # Import modules
 
+import logging
+import sys
+from traceback_with_variables import prints_tb, LoggerAsFile
+
 """
 Import any modules needed here
 """
@@ -9,7 +13,6 @@ Import any modules needed here
 # Import common, logging and settings
 
 import DirectFire.Converter.common as common
-from DirectFire.Converter.logging import logger
 import DirectFire.Converter.settings as settings
 
 # Initialise common functions
@@ -18,10 +21,34 @@ import DirectFire.Converter.settings as settings
 Import any common functions needed here
 """
 
+# Initiate logging
 
-def generate(logger, parsed_data):
+logger = logging.getLogger(__name__)
 
-    logger.log(2, __name__ + ": generator module started")
+
+# Catch exceptions and log
+
+
+@prints_tb(
+    file_=LoggerAsFile(logger),
+    num_context_lines=3,
+    max_value_str_len=9999999,
+    max_exc_str_len=9999999,
+)
+def catch_exception(exc_type, exc_value, exc_trace):
+
+    sys.__excepthook__(exc_type, exc_value, exc_trace)
+
+
+sys.excepthook = catch_exception
+
+
+# Generator
+
+
+def generate(parsed_data):
+
+    logger.info(__name__ + ": generator module started")
 
     # Initialise variables
 
@@ -35,23 +62,23 @@ def generate(logger, parsed_data):
 
     # Generate system
 
-    logger.log(2, __name__ + ": generate system")
+    logger.info(__name__ + ": generate system")
 
     try:
         dst_config.append("hostname " + parsed_data["system"]["hostname"])
     except:
-        logger.log(3, __name__ + ": hostname not found in parsed data")
+        logger.warning(__name__ + ": hostname not found in parsed data")
         pass
 
     try:
         dst_config.append("domain-name " + data["system"]["domain"])
     except:
-        logger.log(3, __name__ + ": domain name not found in parsed data")
+        logger.warning(__name__ + ": domain name not found in parsed data")
         pass
 
     # Generate interfaces
 
-    logger.log(3, __name__ + ": generate interfaces - not yet supported")
+    logger.warning(__name__ + ": generate interfaces - not yet supported")
 
     """
     Generate interfaces
@@ -59,7 +86,7 @@ def generate(logger, parsed_data):
 
     # Generate zones
 
-    logger.log(3, __name__ + ": generate zones - not yet supported")
+    logger.warning(__name__ + ": generate zones - not yet supported")
 
     """
     Generate zones
@@ -67,7 +94,7 @@ def generate(logger, parsed_data):
 
     # Generate static routes
 
-    logger.log(2, __name__ + ": generate static routes")
+    logger.info(__name__ + ": generate static routes")
 
     for route_id, attributes in enumerate(parsed_data["routes"]):
 
@@ -87,7 +114,7 @@ def generate(logger, parsed_data):
 
     # Generate network objects
 
-    logger.log(2, __name__ + ": generate network objects")
+    logger.info(__name__ + ": generate network objects")
 
     for address, attributes in parsed_data["network_objects"].items():
 
@@ -120,7 +147,7 @@ def generate(logger, parsed_data):
 
     # Generate network groups
 
-    logger.log(2, __name__ + ": generate network groups")
+    logger.info(__name__ + ": generate network groups")
 
     for group, attributes in parsed_data["network_groups"].items():
 
@@ -131,7 +158,7 @@ def generate(logger, parsed_data):
 
     # Generate service objects
 
-    logger.log(2, __name__ + ": generate service objects")
+    logger.info(__name__ + ": generate service objects")
 
     for service, attributes in parsed_data["service_objects"].items():
 
@@ -185,7 +212,7 @@ def generate(logger, parsed_data):
 
     # Generate service groups
 
-    logger.log(2, __name__ + ": generate service groups")
+    logger.info(__name__ + ": generate service groups")
 
     for group, attributes in parsed_data["service_groups"].items():
 
@@ -196,7 +223,7 @@ def generate(logger, parsed_data):
 
     # Generate policies
 
-    logger.log(2, __name__ + ": generate policies - not yet supported")
+    logger.info(__name__ + ": generate policies - not yet supported")
 
     # access_lists = {}
 
@@ -339,7 +366,7 @@ def generate(logger, parsed_data):
 
     # Generate NAT
 
-    logger.log(3, __name__ + ": generate NAT - not yet supported")
+    logger.warning(__name__ + ": generate NAT - not yet supported")
 
     """
     Generate NAT policies
@@ -347,6 +374,6 @@ def generate(logger, parsed_data):
 
     # Return generated config
 
-    logger.log(2, __name__ + ": generator module finished")
+    logger.info(__name__ + ": generator module finished")
 
     return dst_config