+++ converter.py
@@ -31,10 +31,15 @@ try:
 except:
     raise ImportError("Could not import module: colorama")
 
+from datetime import datetime
+
+import logging
+
+from traceback_with_variables import activate_by_import
+
 # Import common and settings
 
 import DirectFire.Converter.common as common
-from DirectFire.Converter.logging import logger
 import DirectFire.Converter.settings as settings
 
 # Get arguments
@@ -67,15 +72,30 @@ args = arg_parser.parse_args()
 
 # Initiate logging
 
-logger = logger(args.source, args.destination)
+now = str(datetime.now().strftime("%Y%m%d_%H%M%S"))
 
-logger.log(2, "DirectFire.Converter.main: converter starting")
-logger.log(2, "DirectFire.Converter.main: source format is " + args.source)
+logging.basicConfig(
+    filename="logs/"
+    + now
+    + "_"
+    + str(args.source)
+    + "_"
+    + str(args.destination)
+    + ".log",
+    format="%(asctime)s %(levelname)-8s %(message)s",
+    datefmt="%Y-%m-%d %H:%M:%S",
+    level=logging.DEBUG,
+)
+
+logger = logging.getLogger(__name__)
+
+logger.info("DirectFire.Converter.main: converter starting")
+logger.info("DirectFire.Converter.main: source format is " + args.source)
 
 
 def parse(src_format, src_config, routing_info=""):
 
-    logger.log(2, "DirectFire.Converter.parse: loading parser module for " + src_format)
+    logger.info("DirectFire.Converter.parse: loading parser module for " + src_format)
 
     if src_format == "ciscoasa":  ## Cisco ASA
         from DirectFire.Converter.parsers.ciscoasa import parse
@@ -93,31 +113,29 @@ def parse(src_format, src_config, routing_info=""):
         from DirectFire.Converter.parsers.watchguard import parse
 
     else:
-        logger.log(
-            2,
-            "DirectFire.Converter.parse: failed to load parser module for "
-            + src_format,
+        logger.info(
+            "DirectFire.Converter.parse: failed to load parser module for " + src_format
         )
 
         print(f"{Fore.RED}Error: failed to load parser module.{Style.RESET_ALL}")
 
         exit()
 
-    logger.log(2, "DirectFire.Converter.parse: loaded parser module for " + src_format)
+    logger.info("DirectFire.Converter.parse: loaded parser module for " + src_format)
 
-    logger.log(2, "DirectFire.Converter.parse: starting parse of source configuration")
+    logger.info("DirectFire.Converter.parse: starting parse of source configuration")
 
-    parsed_data = parse(logger, src_config, routing_info)
+    parsed_data = parse(src_config, routing_info)
 
-    logger.log(2, "DirectFire.Converter.parse: completed parse of source configuration")
+    logger.info("DirectFire.Converter.parse: completed parse of source configuration")
 
     return parsed_data
 
 
 def generate(dst_format, parsed_data):
 
-    logger.log(
-        2, "DirectFire.Converter.generate: loading generator module for " + dst_format
+    logger.info(
+        "DirectFire.Converter.generate: loading generator module for " + dst_format
     )
 
     if dst_format == "ciscoasa":  ## Cisco ASA post 8.3
@@ -130,28 +148,27 @@ def generate(dst_format, parsed_data):
         from DirectFire.Converter.generators.fortigate import generate
 
     else:
-        logger.log(
-            2,
+        logger.info(
             "DirectFire.Converter.parse: failed to load generator module for "
-            + dst_format,
+            + dst_format
         )
 
         print(f"{Fore.RED}Error: failed to load generator module.{Style.RESET_ALL}")
 
         exit()
 
-    logger.log(
-        2, "DirectFire.Converter.generate: loaded generator module for " + dst_format
+    logger.info(
+        "DirectFire.Converter.generate: loaded generator module for " + dst_format
     )
 
-    logger.log(
-        2, "DirectFire.Converter.generate: starting generation of destination output"
+    logger.info(
+        "DirectFire.Converter.generate: starting generation of destination output"
     )
 
-    dst_config = generate(logger, parsed_data)
+    dst_config = generate(parsed_data)
 
-    logger.log(
-        2, "DirectFire.Converter.generate: completed generation of destination output"
+    logger.info(
+        "DirectFire.Converter.generate: completed generation of destination output"
     )
 
     return dst_config
@@ -161,8 +178,8 @@ def main(src_format, dst_format, routing_info=""):
 
     # Load source configuration file
 
-    logger.log(
-        2, "DirectFire.Converter.main: loading source configuration from " + args.config
+    logger.info(
+        "DirectFire.Converter.main: loading source configuration from " + args.config
     )
 
     try:
@@ -172,10 +189,9 @@ def main(src_format, dst_format, routing_info=""):
 
     except:
 
-        logger.log(
-            4,
+        logger.error(
             "DirectFire.Converter.main: source file either not found or not readable "
-            + args.config,
+            + args.config
         )
 
         print(
@@ -193,10 +209,9 @@ def main(src_format, dst_format, routing_info=""):
 
         except:
 
-            logger.log(
-                4,
+            logger.error(
                 "DirectFire.Converter.main: routing file either not found or not readable "
-                + args.config,
+                + args.config
             )
 
             print(
@@ -207,17 +222,17 @@ def main(src_format, dst_format, routing_info=""):
 
     # Run configuration parser
 
-    logger.log(2, "DirectFire.Converter.main: running configuration parser")
+    logger.info("DirectFire.Converter.main: running configuration parser")
 
     parsed_data = parse(
         src_format=src_format, src_config=src_config, routing_info=routing_info
     )
 
-    logger.log(2, "DirectFire.Converter.main: configuration parser finished")
+    logger.info("DirectFire.Converter.main: configuration parser finished")
 
     # Output
 
-    logger.log(2, "DirectFire.Converter.main: running configuration generator")
+    logger.info("DirectFire.Converter.main: running configuration generator")
 
     dst_config = generate(dst_format=dst_format, parsed_data=parsed_data)
 
@@ -226,9 +241,9 @@ def main(src_format, dst_format, routing_info=""):
 
     ### add support for output to file
 
-    logger.log(2, "DirectFire.Converter.main: configuration generator finished")
+    logger.info("DirectFire.Converter.main: configuration generator finished")
 
-    logger.log(2, "DirectFire.Converter.main: converter exiting")
+    logger.info("DirectFire.Converter.main: converter exiting")
 
 
 if __name__ == "__main__":